requirements.txt
streamlit
pandas
opencv-python-headless
google-genai
altair
fpdf    